public-prediction/requirements.txt
kafka-python
langchain
openai
langchain-openai
python-dotenv